Backend optimizations and new features
Optimizations:
- FetchType.LAZY on all entity relationships
- JOIN FETCH queries for N+1 prevention
- Database indexes on frequently queried columns
- HikariCP connection pooling configuration
- Caching for PlatformSettings
- COUNT/SUM queries instead of loading full entities
- Rate limiting with Bucket4j
- Global exception handling with standardized API errors

New Features:
- Admin wallet, analytics, billboard management
- Owner wallet, bank, dashboard, notifications
- Payment splits and payout services
- Two-factor authentication enhancements
- Security controllers and services
- Smart pricing and tax services

Bug Fixes:
- Fixed LAZY loading serialization with Hibernate6Module
- Fixed missing endpoint 404 handling
- Added @JsonIgnoreProperties for proxy serialization

+++ src/main/java/com/billboarding/DTO/Advertiser/CampaignAnalyticsDTO.java
@@ -1,9 +1,11 @@
 package com.billboarding.DTO.Advertiser;
+
 import lombok.Builder;
 import lombok.Data;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
+
 @Data
 @Builder
 public class CampaignAnalyticsDTO {
@@ -12,13 +14,12 @@ public class CampaignAnalyticsDTO {
     private String campaignName;
     private String status;
 
-    private BigDecimal budget;
+    private Double budget;
     private BigDecimal spent;
 
     private Long impressions;
-
-    private Double cpm;                 // cost per 1000 impressions
-    private Double budgetUtilization;   // %
+    private Double cpm;
+    private Double budgetUtilization;
 
     private LocalDate startDate;
     private LocalDate endDate;
